requirements.txt
asgiref==3.11.1
attrs==25.4.0
bleach==6.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
webencodings==0.5.1
    